feat: add types to profile-feed component
